fold setup-message-repo.ts into conversation-store.ts + import cleanup
